--- a/spec/models/miq_product_feature_spec.rb
+++ b/spec/models/miq_product_feature_spec.rb
@@ -1,4 +1,6 @@
 require "spec_helper"
+require 'tmpdir'
+require 'pathname'
 
 describe MiqProductFeature do
   before do
@@ -31,6 +33,46 @@ describe MiqProductFeature do
       expect { deleted.reload }.to raise_error(ActiveRecord::RecordNotFound)
       changed.reload.name.should == "About"
       unchanged.reload.updated_at.should be_same_time_as unchanged_orig_updated_at
+    end
+  end
+
+  context ".seed_features" do
+    let(:feature_path) { Pathname.new(Dir.tmpdir).join(rand(1000).to_s) }
+
+    before do
+      FileUtils.mkdir_p(feature_path)
+    end
+
+    after do
+      FileUtils.rm_rf(feature_path)
+    end
+
+    it "additional yaml feature" do
+      base = {
+        :feature_type => "node",
+        :identifier   => "everything",
+        :children     => [
+          {
+            :feature_type => "node",
+            :identifier   => "one",
+            :children     => []
+          }
+        ]
+      }
+
+      additional = {
+        :feature_type => "node",
+        :identifier   => "two",
+        :children     => []
+      }
+
+      additional_file = feature_path.join("additional.yml")
+      File.write("#{feature_path}.yml", YAML.dump(base))
+      File.write(additional_file, YAML.dump(additional))
+
+      status_seed = MiqProductFeature.seed_features(feature_path)
+      MiqProductFeature.count.should eq(3)
+      expect(status_seed[:created]).to match_array %w(everything one two)
     end
   end
 end