We have several duplicated identifiers in the `db/fixtures/miq_product_features.yml` when they must be unique.  The seed method will create on the first instance and update on each subsequent instance and the last one in the file always wins.  Since only the last one survives, I chose to delete these "all but last duplicates" cc @h-kataria @dclarizio 

Added a missing test for "providing additional product features used by the UI plugins" (from #1680)

Changed the tests to not run the seed against the enormous miq_product_features.yml file except when we try to detect duplicate identifiers in changes to this file.

Tests before:
`Finished in 10.4 seconds`

Tests after:
`Finished in 3.25 seconds`

MiqProductFeature.seed is also a little faster:
Initial seed:            3 s -> 2.3 s
Subsequent seed:  0.8 s -> 0.2 s