@Fryguy 2015-09-30
If identifier is meant to be unique...let's put a unique validation on the model.  Then you don't need to self-check.  Seed will just straight-up fail if there are duplicates.