@jrafanie 2015-09-30
> Finally, I also removed the unused "feature" to seed product features from the non-existing db/fixtures/miq_product_features directory.
> I believe this is for downstream productization. Can you verify?

I looked, I couldn't see anything besides an ae datastore... @h-kataria @dclarizio can you confirm?