@jrafanie 2015-09-30
> If identifier is meant to be unique...let's put a unique validation on the model. Then you don't need to self-check. Seed will just straight-up fail if there are duplicates.

It's already there.  Seed will create or update based on the identifier.  It was looping through each identifier section of the yaml, fetch the row (N+1) matching the record, creating or updating it.

For example, tail of a log which no product features + seeding showing that we had duplicate identifiers in the yaml that were first created, then updated:
```
[----] I, [2015-09-30T12:52:04.120682 #32449:3fd8b5460208]  INFO -- : MIQ(MiqProductFeature.seed_from_hash) Creating product feature: Identifier: [ab_group_new], Name: [Add a new Button Group]
[----] I, [2015-09-30T12:52:05.439592 #32449:3fd8b5460208]  INFO -- : MIQ(MiqProductFeature.seed_from_hash) Updating product feature: Identifier: [ab_group_new], Name: [Add Button Group]
```