@dclarizio 2015-09-30
> The other thing it might be used for is the UI module plugin piece.

@martinpovolny?