@jrafanie 2015-09-30
@Fryguy I think I applied all of your suggestions