@jrafanie 2015-10-06
> Finally, I also removed the unused "feature" to seed product features from the non-existing db/fixtures/miq_product_features directory. cc @Fryguy @gtanzillo @dclarizio Please let me know if this was wrong, I can't see this being used.

Ok, rebased, @martinpovolny any idea about the above directory that doesn't exist anywhere?  I'd like to remove that feature if we're not using it.  Thanks!