@jrafanie 2015-10-14
@martinpovolny @dclarizio Can you review/merge this? Thanks!