@h-kataria 2015-10-14
@jzigmund can you look over changes in db/fixtures/miq_product_features.yml, i think you are renaming a feature in your PR https://github.com/ManageIQ/manageiq/pull/3987 that's being deleted by this PR.