@jrafanie 2015-10-22
@h-kataria @jzigmund @martinpovolny Can we get this reviewed?  Part of the benefit of the revised spec is that it will now fail if we accidentally introduce a new duplicate identifier.  Currently, we have duplicate identifiers we had no idea we had, nor if there are parts of the UI that needed the now deleted duplicates, and finally, it was costly both in tests and in seeding to create and then update the same duplicated identifier as only the last one in "wins".