@jrafanie 2015-11-02
@h-kataria @martinpovolny @jzigmund @himdel @skateman Can someone please review the `Finally, I also removed the unused "feature" to seed product features from the non-existing db/fixtures/miq_product_features directory` part of this PR?   Do we ever seed product features anywhere using a yaml file that is not db/fixtures/product_features.yml?  That feature is being removed in this PR since it's unused AFAIK.