@martinpovolny 2015-11-03
@jrafanie : The possibility to add product features is needed for the UI plugins to work. See https://github.com/ManageIQ/manageiq/pull/1680. We need to provide a different way for the plugins to add product features if we remove the `db/fixtures/miq_product_features directory`.