@jrafanie 2015-11-03
Thanks @martinpovolny, I'll fix this PR to not kill that feature since it's in use.