@martinpovolny 2015-11-03
@jrafanie : @Fryguy did not like the way I hacked this originally. So we might use the opportunity to address his concerns.