@Fryguy 2015-11-05
@martinpovolny Yeah, we can revisit, but not for this performance PR.