@jrafanie 2015-11-06
Ok @martinpovolny @Fryguy, I kept the UI plugins feature of loading additional product features and added a test.  I also sped up the tests some more.  See updated description above.