@chessbyte 2015-11-15
@Fryguy @martinpovolny please review